+++ NEWS.md
@@ -53,6 +53,7 @@ A major refactor went into the codebase to cleanup several leaking processes. Th
 - Always recalculate screen space location of the control center and info center
 - When a faulty plugin is loaded you now get the correct name in the error message
 - Fetch wifi list now works when no connection is currently active
+- Right bar now fits correctly when no top-panel is present
 
 ### Dev Notes
 - The inputfield now allows editing of the field